Refactor fedbiomed_gui/server into fedbiomed/restful
