+++ fedbiomed/restful/wsgi.py
@@ -1,4 +1,5 @@
 from .application import app
+from .config import config
 
 # Production
 if __name__ == "__main__":
